backend/requirements.txt
djangorestframework
djangorestframework-gis
django-cors-headers
django-phonenumber-field
django-nested-admin
django-admin-rangefilter
phonenumbers
Django
gunicorn
Pillow
psycopg2-binary
sentry-sdk
pytest-django
pytest
python-dotenv
pyyaml
django-debug-toolbar
requests
simplejson